refactor: 절대 경로 변경
