+++ src/pages/form/FormCreatePage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as S from 'pages/form/FormCreatePage.style';
-import { FormProvider } from './context/FormContext';
-import FormSubmitter from './components/FormSubmitter';
+import { FormProvider } from '@Form/context/FormContext';
+import FormSubmitter from '@Form/components/FormSubmitter';
 
 function FormCreate() {
   return (
