+++ src/pages/rolling/paper/components/RollingPaperItem.jsx
@@ -9,15 +9,6 @@ function RollingPaperItem({ data, onClickModal, getPaperData }) {
   };
   return (
     <CardList onClick={handleOpenModal}>
-      {/* <Card
-        key={data.id}
-        name={data.sender}
-        profileImg={data.profileImageURL}
-        label={data.relationship}
-        text={data.content}
-        date={data.createdAt}
-      /> */}
-
       <Card key={data.id} data={data} />
     </CardList>
   );
